[EI-200] [web] refactor: Refactor import paths to use the new indicator service module
