+++ web/app/business/services/view-model/indicator-list/indicators/cryptocurrency-indicator.service.ts
@@ -1,5 +1,5 @@
 import { CryptocurrenciesIndicatorResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
-import { Indicator } from '../indicator-view-model.service';
+import { Indicator } from './indicator.service';
 
 export class CryptocurrencyIndicator extends Indicator {
   readonly symbol: String;
