+++ web/app/business/services/view-model/indicator-list/indicators/etf-indicator.service.ts
@@ -1,5 +1,5 @@
 import { EtfIndicatorResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
-import { Indicator } from '../indicator-view-model.service';
+import { Indicator } from './indicator.service';
 
 export class EtfIndicator extends Indicator {
   readonly symbol: String;
