+++ web/app/business/services/view-model/indicator-list/indicators/forex-pair-indicator.service.ts
@@ -1,4 +1,4 @@
-import { Indicator } from '../indicator-view-model.service';
+import { Indicator } from './indicator.service';
 
 import { ForexPairsIndicatorResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
