+++ web/app/business/services/view-model/indicator-list/indicators/fund-indicator.service.ts
@@ -1,5 +1,5 @@
 import { FundsIndicatorResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
-import { Indicator } from '../indicator-view-model.service';
+import { Indicator } from './indicator.service';
 
 export class FundIndicator extends Indicator {
   readonly symbol: String;
