+++ web/app/business/services/view-model/indicator-list/indicators/indicator.service.ts
@@ -1,0 +1,17 @@
+import { IndicatorType } from '@/app/store/stores/numerical-guidance/indicator-list.store';
+
+export abstract class Indicator {
+  readonly id: String;
+  readonly indicatorType: IndicatorType;
+
+  constructor(id: String, indicatorType: IndicatorType) {
+    this.id = id;
+    this.indicatorType = indicatorType;
+  }
+
+  abstract get symbol(): String;
+
+  abstract get name(): String;
+
+  abstract get exchange(): String;
+}
